Move SchedulerInvocation to a new file. No functional changes.
